requirements.txt
beautifulsoup4==4.6.0
python_Levenshtein==0.21.1
Requests==2.31.0
